Stop Settings from constructing inactive panes on open
Profiling showed the first Settings presentation was dominated by main-thread SwiftUI construction of every settings pane plus pane-local side effects such as browser import detection, workspace color palette reloads, and shortcut recorder row initialization. The detail now renders only the selected top-level pane, applies pending Settings navigation before the first detail render, preserves dirty Browser allowlist drafts across section changes, and moves browser detection off the main actor.

Constraint: The fix must stay surgical and avoid sleeps, new dependencies, broad Settings refactors, or Swift file-length budget growth.

Rejected: Pre-warming every pane after launch | keeps the expensive work and risks moving the stall into startup.

Rejected: Off-main shell/disk scan cleanup alone | profiling showed the dominant cost was eager main-thread view construction.

Confidence: medium

Scope-risk: moderate

Directive: Keep Settings search/navigation selecting the target section before requesting an anchor scroll; hidden panes do not have live anchors.

Tested: ./scripts/reload.sh --tag issue-3384-settings-slow-open --launch succeeded and launched the tagged app.

Not-tested: Local tests by instruction; CI owns tests_v2/test_settings_open_latency.py.
